Add files via upload
It's just a version 1. I didn't able to make events, because i've had many errors. So let's see this version, if you can help me it'll a great pleasure. Please
you'll see that the file index of the pageweb is one that have php code.
i'll continue to work it on